Added the `deps` field to file nodes.
Currently, each dep is a resolved import path
relative to the project dir, e.g. /src/foo/bar.ts.
The target file node can be found by looking
for a dir node with name="src", then a dir node
with name="foo" and finally for a file node with
name="bar.ts". Eventually, each node will have
an id and the deps will be a list of such ids.